refactor(test): extract shared PHPStan fixture runner and fix review findings
