+++ examples/Actions/User/BulkOperationsDemo.php
@@ -5,6 +5,7 @@ declare(strict_types = 1);
 namespace Pekral\Arch\Examples\Actions\User;
 
 use Pekral\Arch\Action\ArchAction;
+use Pekral\Arch\Examples\Services\User\UserModelManager;
 use Pekral\Arch\Examples\Services\User\UserModelService;
 
 /**
@@ -13,8 +14,7 @@ use Pekral\Arch\Examples\Services\User\UserModelService;
 final readonly class BulkOperationsDemo implements ArchAction
 {
 
-    public function __construct(private readonly UserModelService $userModelService)
-    {
+    public function __construct(private UserModelService $userModelService, private UserModelManager $userModelManager,) {
     }
 
     /**
@@ -42,7 +42,7 @@ final readonly class BulkOperationsDemo implements ArchAction
             ['name' => 'Eve Davis', 'email' => 'eve@example.com', 'password' => 'password000'],
         ];
 
-        $this->userModelService->getModelManager()->insertOrIgnore($mixedUsers);
+        $this->userModelManager->insertOrIgnore($mixedUsers);
 
         $bulkUpdateResult = $this->userModelService->bulkUpdate($updateData);
